requirements.txt
streamlit==1.39.0
pymongo==4.8.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
plotly==5.24.1
seaborn==0.13.2
matplotlib==3.9.2
python-dotenv==1.0.1
joblib==1.4.2
tqdm==4.66.5
requests==2.32.3
